+++ OlympGuide.Application/Features/TestData/ITestDataService.cs
@@ -1,0 +1,8 @@
+﻿namespace OlympGuide.Application.Features.TestData
+{
+    public interface ITestDataService
+    {
+        Task<int> CreateTestData();
+        Task<int> DeleteTestData();
+    }
+}
