+++ OlympGuide.Application/ServiceRegistration.cs
@@ -2,6 +2,7 @@
 using OlympGuide.Application.Features.Reservation;
 using OlympGuide.Application.Features.SportField;
 using OlympGuide.Application.Features.SportFieldProposal;
+using OlympGuide.Application.Features.TestData;
 using OlympGuide.Application.Features.User;
 using OlympGuide.Domain.Features.User;
 
@@ -15,6 +16,7 @@ namespace OlympGuide.Application
             services.AddScoped<ISportFieldProposalService, SportFieldProposalService>();
             services.AddScoped<IReservationService, ReservationService>();
             services.AddScoped<IUserService, UserService>();
+            services.AddScoped<ITestDataService, TestDataService>();
         }
     }
 }
